image_to_text/app/requirements.txt
Flask==2.1.2
Pillow==8.3.1
pytesseract==0.3.8
